+++ localization/el.js
@@ -1,5 +1,5 @@
 /**
- * jui_filter_rules simple localization - GREEK
+ * jui_filter_rules v1.0.5 simple localization - GREEK
  *
  * DO NOT CHANGE this file, as it will be overwritten in next update.
  * To use different values, write and use a similar structure js file.
@@ -53,5 +53,3 @@ var rsc_jui_fr = {
     error_invalid_datetime: "Η ημερομηνία δεν είναι σωστή",
     error_converting_value: "Λάθος κατά τη μετατροπή της τιμής που δόθηκε"
 };
-
-