+++ localization/en.js
@@ -1,5 +1,5 @@
 /**
- * jui_filter_rules simple localization - ENGLISH
+ * jui_filter_rules v1.0.5 simple localization - ENGLISH
  *
  * DO NOT CHANGE this file, as it will be overwritten in next update.
  * To use different values, write and use a similar structure js file.
@@ -53,4 +53,3 @@ var rsc_jui_fr = {
     error_invalid_datetime: "Invalid date format",
     error_converting_value: "Error converting value"
 };
-